requirements.txt
chardet==3.0.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
colorama==0.4.4
colorgram.py==1.2.0
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.2
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
psycopg2-binary==2.9.2
PyJWT==2.1.0
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
twilio==7.0.0
urllib3==1.26.7
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3

